Move PInvoke wrapper methods to PInvoke/Modules/ + Add constant definition file
