+++ TeamDEV.Asl/Const.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+namespace TeamDEV.Asl {
+    static class Const {
+        public const string UnknownString = "Unknown";
+    }
+}
