+++ TeamDEV.Asl/PInvoke/Modules/Kernel32.cs
@@ -3,11 +3,10 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Text;
 
-using TeamDEV.Asl.PInvoke.Internal.Enumerations;
 using TeamDEV.Asl.PInvoke.Enumerations;
 using TeamDEV.Asl.PInvoke.Structures;
 
-namespace TeamDEV.Asl.PInvoke.Internal.Methods {
+namespace TeamDEV.Asl.PInvoke.Modules {
     /// <summary>
     /// 
     /// </summary>
